requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.2
h11==0.14.0
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
psutil==5.9.6
py-cpuinfo==9.0.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.0.227
urllib3==2.1.0
uvicorn==0.24.0.post1
yarl==1.9.4
